test: Add tests for auth reducer
