+++ ui/package.json
@@ -49,7 +49,8 @@
     "@babel/preset-env": "7.13.8",
     "@babel/preset-react": "^7.10.4",
     "@testing-library/jest-dom": "^5.11.5",
-    "@testing-library/react": "^11.1.1",
+    "@testing-library/react": "^11.2.6",
+    "@testing-library/react-hooks": "^5.1.1",
     "@types/jest": "^26.0.15",
     "@types/leaflet": "^1.5.19",
     "@types/node": "^14.14.6",
