+++ ui/src/hooks/auth/index.tsx
@@ -6,30 +6,10 @@ import {
   useGoogleLogout,
 } from "react-google-login";
 
+import { FailureState, INIT_STATE, reducer,State } from "./reducer";
+
 const CLIENT_ID =
   "722954318269-211qsag71c3bsfjik321h9sa0kmbnelf.apps.googleusercontent.com";
-
-export enum FailureState {
-  InvalidEmail,
-  PopupFail,
-  Unknown,
-}
-
-type State = {
-  loading: boolean;
-  auth?: GoogleLoginResponse;
-  failure?: FailureState;
-};
-
-type Action =
-  | { type: "START_LOADING" }
-  | {
-      type: "SUCCESS";
-      response: GoogleLoginResponse;
-    }
-  | { type: "FAILURE"; failure: FailureState }
-  | { type: "LOGOUT_SUCCESS" }
-  | { type: "LOGOUT_FAILURE" };
 
 type AuthContextType = State & {
   googleLoaded: boolean;
@@ -42,48 +22,6 @@ const AuthContext = React.createContext({});
 
 export function useAuth(): AuthContextType {
   return React.useContext(AuthContext) as AuthContextType;
-}
-
-const INIT_STATE: State = Object.freeze({
-  loading: false,
-  auth: undefined,
-  failure: undefined,
-});
-
-function reducer(state: State, action: Action): State {
-  switch (action.type) {
-    case "START_LOADING": {
-      return {
-        ...state,
-        loading: true,
-      };
-    }
-    case "SUCCESS": {
-      return {
-        ...state,
-        loading: false,
-        failure: undefined,
-        auth: action.response,
-      };
-    }
-    case "FAILURE": {
-      return {
-        ...state,
-        loading: false,
-        failure: action.failure,
-        auth: undefined,
-      };
-    }
-    case "LOGOUT_SUCCESS": {
-      return {
-        ...state,
-        failure: undefined,
-        auth: undefined,
-      };
-    }
-    default:
-      return state;
-  }
 }
 
 export function useProvideAuth(): AuthContextType {
